==> R/04_description_av.rmd <==
---
title: 'Analysis of the virtual assistants in the eligible papers included in the scoping review'
author: 
  - name: "Carlos Granell"
    affiliation: Universitat Jaume I, Spain
date: "`r format(Sys.time(), '%d %B, %Y')`"
abstract: |
  This analysis generates tables and figures related to the characteristics 
  of the virtual assistants included in the eligible papers.
  It is a collaborative research activity as part of the [IDEAIS project](http://www.redideais.net/).

output:
  html_document:
    df_print: paged
    toc: yes
    toc_float: true
    toc_depth: 3
    number_sections: true
    theme: readable
  word_document:
    toc: yes
urlcolor: blue
---


```{r setup, include=FALSE, eval=TRUE}
knitr::opts_chunk$set(echo = FALSE)
```


# Required libraries

```{r load_libraries, echo=TRUE, message=FALSE, warning=FALSE}
library(here)
library(tidyverse)
library(kableExtra)
library(DT)
library(stringr)
library(grid)
library(gridExtra)

library(ggforce) # for parallel sets diagram
library(gghighlight)

# just in case
set.seed(nchar("IDEAIS - scoping review paper"))
```

# Data 

```{r load_data, warning=FALSE, comment=FALSE}
papers2010_2019 <- readRDS(file = here("data", "papers2010_2019.rda"))
data_av1 <- readRDS(file = here("data", "data_av1.rda"))
data_av2 <- readRDS(file = here("data", "data_av2.rda"))
data_geo <- readRDS(file = here("data", "data_geo.rda"))


cols(
  group = col_character(),
  domain = col_character(),
  geo = col_character(),
  desc = col_character()
) -> export_cols


read_csv2(
  file = here::here("data", "data_sankey.csv"), 
  col_types = export_cols
) %>% 
  select(group, domain, geo) %>%
  group_by(group, domain, geo) %>%
  summarise(freq = n()) -> sankey 


```

```{r all_data_join}
all_data <- 
  papers2010_2019 %>%
  inner_join(data_av1, by="id") %>%
  inner_join(data_geo, by="id") %>%
  select(id, year, av_app_norm, av_application, av_name, av_place, av_status, av_arch, av_platform,
         geo_goal_norm, geo_goal, geo_sources, geo_tech, geo_inputmode, geo_outputmode, geo_standards, 
         geo_process, geo_processtype)

```


```{r all_data_table}
all_data %>%
  select(ID = id,
         Year = year,
         `[AV] Application` = av_application,
         `[AV] Name` = av_name,
         `[AV] Place`= av_place,
         `[AV] Dev. status` = av_status,
         `[AV] Arch.` = av_arch,
         `[AV] Platform` = av_platform,
         `[Geo] Goal` = geo_goal,
         `[Geo] Sources` = geo_sources,
         `[Geo] Tech` = geo_tech,
         `[Geo] Input` =geo_inputmode,
         `[Geo] Output` = geo_outputmode,
         `[Geo] Standards?` = geo_standards,
         `[Geo] Process?` = geo_process,
         `[Geo] Process type` = geo_processtype) %>%
  arrange(desc(Year)) %>%
  datatable(rownames = FALSE,
            filter = "top",
            class = "table-bordered table-condensed hover",
            extensions = c("Buttons"),
            options = list(
              pageLength = 10, #autoWidth = TRUE,
              dom = 'Blfrtip',  # https://datatables.net/reference/option/dom
              buttons = c('copy', 'csv', 'excel', 'pdf', 'print')
            )
  )
```



# Analysis

The following tables and figures are the results of the analysis of the characteristics of the virtual assistants.

## Figure 4. Interaction between categorical variables (group, application domain and geospatial use) of virtual assistants

```{r parallel_set_plot, comment=FALSE, dpi=600, fig.height=14, fig.width=14}

# Source: https://rviews.rstudio.com/2019/09/19/intro-to-ggforce/
# Source: https://cran.r-project.org/web/packages/ggforce/ggforce.pdf
sankey %>%
  rename(`Application domain/\nFeature or aspect` = domain,
         `Geospatial usage` = geo,
         `Group` = group) %>%
  gather_set_data(1:3) %>% 
  ggplot(aes(x, id = id, split = y, value = freq)) +
  geom_parallel_sets(aes(fill = Group), alpha = 0.3, axis.width = 0.2, show.legend = FALSE) +
  geom_parallel_sets_axes(axis.width = 0.3, color = "lightgrey", fill = "white") +
  geom_parallel_sets_labels(colour = "black", angle = 0, size = 4) +
  scale_fill_manual(values=c("#E69F00", "#999999")) +
  theme_bw() +
  theme(panel.grid.minor = element_blank()) +
  theme(panel.grid.major = element_blank()) +
  theme(axis.title.x = element_blank()) +
  theme(axis.title.y = element_blank(), axis.text.y = element_blank(), axis.ticks.y = element_blank())
  

ggsave(here("figs", "fig04_parallelsets.png"))

```

## Useful information/stats 

Tables below show the numbers and percentages used in Section 3 of the paper. They are not shown as tables per se but as pieces of information intertwined in the narrative. 

```{r av_stats}
papers <- 
  papers2010_2019 %>% 
  inner_join(data_av1, by="id") %>%
  arrange(id)

n_papers <- nrow(papers)

```


_Temporal distribution of virtual assistants according to their development status (`av_status`)_

```{r dev_status}
papers %>%
  group_by(av_status) %>% 
  summarise(n = n()) %>%
  mutate(proportion  = n / n_papers, 
         proportion_lbl = paste0(round(proportion*100,0), "%")) %>%
  arrange(n) %>%
  select(`Development stage of virtual assisstants` = av_status,
         `N` = n,
         `%`= proportion_lbl) %>%
  knitr::kable(format="html", escape = T, booktabs = TRUE) %>%
  kableExtra::kable_styling(bootstrap_options = "striped", full_width = FALSE) %>%
  kableExtra::scroll_box(height = "300px")
  
```


_What AVs do they reach the status of operational implementation?_
```{r dev_status_operational}

papers %>%
  filter(av_status=="Implementation") %>%
  select(id, av_name) %>%
  knitr::kable(caption = "AVs that reach operational implementation",
               format="html", escape = T, booktabs = TRUE) %>%
  kableExtra::kable_styling(bootstrap_options = "striped", full_width = FALSE) %>%
  kableExtra::scroll_box(height = "300px")

```



_What are the most used platforms?_

```{r platforms}
platforms <- 
  papers %>%
  select(av_platform) %>%
  group_by(av_platform) %>%
  summarise(n = n()) %>%
  arrange(desc(n))


no_platforms <- nrow(platforms %>% filter(!is.na(av_platform)))

platforms %>%
  knitr::kable(caption = "Types of platforms",
               format="html", escape = T, booktabs = TRUE) %>%
  kableExtra::kable_styling(bootstrap_options = "striped", full_width = FALSE) %>%
  kableExtra::scroll_box(height = "300px")
```


Out of `r n_papers` papers, the implementated platform is mentioned in `r no_platforms` studies, which is about `r round(no_platforms/n_papers*100, 0)`%.


_How many virtual assisstants are named?_

```{r named}
named <- 
  papers %>%
  select(id, av_name, year, av_app_norm, av_application) %>%
  filter(!is.na(av_name))

  
named %>%
  knitr::kable(caption = "Named virtual assisstants",
               format="html", escape = T, booktabs = TRUE) %>%
  kableExtra::kable_styling(bootstrap_options = "striped", full_width = FALSE) %>%
  kableExtra::scroll_box(height = "300px")
```

Number of named virtual assisstants `r round(nrow(named)/n_papers*100,0)`% (`r nrow(named)`/`r n_papers`)


## Figure 5. Temporal distribution of virtual assistants according to their technical approach (`av_arch`) and development status (`av_status`)


```{r fig5_facet, warning=FALSE, comment=FALSE, dpi=600, fig.height=8, fig.width=12}
papers %>%
  select(year, av_arch, av_status) %>%
  group_by(year, av_arch, av_status) %>%
  summarise(n = n()) %>%
  ggplot(aes(x=year, y=av_arch, color=av_status, group=av_status)) +
  geom_point(aes(size=n*4),  alpha=.5, na.rm = TRUE) +
  geom_text(aes(label=n), color="white", size=4, na.rm = TRUE) +
  scale_x_continuous(breaks = seq(2010,2020, by=1)) +
  scale_size_area(max_size=15) +
  scale_color_brewer(palette = "Dark2") +
  labs(title="Temporal distribution of studies \naccording to their technical approach and development status",
       x="Year", 
       y="Technical approach",
       caption="Source: authors") +
  facet_wrap(~av_status, ncol = 2, strip.position = "right") +
  # Which legend to show
  guides(color="none", size = "none") +
  theme_bw() + 
  theme(panel.grid.minor = element_blank()) +
  theme(panel.background = element_blank()) +
  theme(strip.text.y = element_text(size=10, face="bold"),
        strip.background = element_rect(colour="lightgray", fill="gray90"))

ggsave(here("figs", "fig05-bubbles_facet.png"))
```


## Table 2. Summary of the virtual assisstants

```{r longtable_data}
app_domains = c("Disaster management", 
                "Education", 
                "Education; Entertaintment", 
                "Healthcare",
                "Recommendation", 
                "Tourism", 
                "Tourism; Cultural heritage", 
                "Tourism; Entertaintment")
others = c("Conversation", 
           "Linguistic", 
           "Navigation")

table_researchareas <-
  data_av2 %>%
  group_by(id) %>% 
  arrange(av_researchareas) %>%
  summarise(av_areas  = paste(stringr::str_to_sentence(av_researchareas), collapse =", ")) %>%
  arrange(id)

  
table_papers <- 
  papers2010_2019 %>%
  inner_join(data_av1, by="id") %>%
  inner_join(data_geo, by="id") %>%
  inner_join(table_researchareas, by="id") %>%
  select(av_app_norm, geo_goal_norm, year, id, av_application, geo_goal, av_areas) %>%
  mutate(ref = paste0("[", id, "] ", year),
         group = ifelse(av_app_norm %in% app_domains, 1, 2)) %>%
  arrange(group, av_app_norm, year)

```

```{r longtable_output}
table_papers  %>%
  select(`G`= group,
         `App domain`= av_app_norm,
         `Geo goal`= geo_goal_norm,
         `Ref - Year` = ref,
         `App desc` = av_application,
         `Geo desc` = geo_goal,
         `Research areas`= av_areas) %>%
  datatable(rownames = FALSE,
            filter = "top",
            class = "table-bordered table-condensed hover",
            extensions = c("Buttons"),
            options = list(
              pageLength = 10, #autoWidth = TRUE,
              dom = 'Blfrtip',  # https://datatables.net/reference/option/dom
              buttons = c('copy', 'csv', 'excel', 'pdf', 'print')
            )
  )


table_papers  %>%
  select(`G`= group,
         `App domain`= av_app_norm,
         `Geo goal`= geo_goal_norm,
         `Ref - Year` = ref,
         `App desc` = av_application,
         `Geo desc` = geo_goal,
         `Research areas`= av_areas) %>%
  knitr::kable(format="html", escape = T, booktabs = TRUE) %>%
  kable_styling(bootstrap_options = c("striped", "hover", "condensed", "responsive")) %>%
  kableExtra::collapse_rows(columns = 1:2, valign = "top") %>%
  kableExtra::scroll_box(height = "300px")
```



## Figure 6. Most influential research areas found in the reviews papers

```{r research_areas, warning=FALSE, comment=FALSE}
researchareas <-
  table_researchareas %>%
  separate_rows(av_areas, sep = ",") %>%
  mutate(av_areas = stringr::str_trim(av_areas)) %>%
  group_by(av_areas) %>%
  tally(sort = TRUE)


ra_greater_one <-
  researchareas %>%
  filter(n > 1)

ra_equal_one <-
  researchareas %>%
  filter(n == 1) 

no_ra_equal_one <-  
  ra_equal_one %>% count() %>% .$n

ra_greater_one <-
  ra_greater_one %>%
  tibble::add_row(av_areas = "Others", n = no_ra_equal_one) %>%
  tibble::add_column(place = c(1:nrow(.)), .before = 0)

total <- sum(ra_greater_one$n)

ra_greater_one <-
  ra_greater_one %>% 
  mutate(per = round(n / total*100,0))


# Ordering data by "place" in decreasing order, to keep "Others" the last bar
ra_greater_one$av_areas <- 
  factor(ra_greater_one$av_areas, 
         levels = ra_greater_one$av_areas[order(ra_greater_one$place, decreasing = TRUE)])


```


```{r research_areas_plot, comment=FALSE, warning=FALSE, dpi=600, fig.height=8, fig.width=10}

plot <- 
  ra_greater_one %>%
  ggplot(aes(x = av_areas, y = per)) + 
  geom_bar(stat = "identity") + 
  geom_text(aes(label = paste0(per, "%")), hjust = -.1, size=3) +
  labs(
    title="Percentages of the research areas identified in the reviewed papers.",
    x = "Research areas", 
    y = "Percentage",
    caption="Source: authors") +
  coord_flip() +
  gghighlight(av_areas == "Others") +
  theme_minimal() +
  theme(panel.grid.minor.y = element_blank()) +
  theme(panel.grid.major.y = element_blank())
  

ggsave(plot, filename = here::here("figs", "fig06.png"))
```


```{r research_areas_include_plot, dpi=72}
knitr::include_graphics(here::here("figs", "fig06.png"))
```



```{r researchareas_one_table}


ra_equal_one %>%
  select(av_areas) %>%
  rename(`Others: research areas cited once` = av_areas) %>%
  knitr::kable(format="html", escape = T, booktabs = TRUE) %>%
  kable_styling(bootstrap_options = c("striped")) %>%
  kableExtra::scroll_box(height = "300px")

```

# Runtime environment description


```{r session_info, echo=FALSE}
devtools::session_info(include_base = TRUE)
```
